+++ jeap-messaging-sequential-inbox/src/main/java/ch/admin/bit/jeap/messaging/sequentialinbox/jpa/SpringDataJpaBufferedMessageRepository.java
@@ -6,14 +6,8 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.time.ZonedDateTime;
-
 @Repository
 interface SpringDataJpaBufferedMessageRepository extends JpaRepository<BufferedMessage, Long> {
-
-    @Modifying
-    @Query(nativeQuery = true, value = "DELETE FROM buffered_message WHERE sequence_instance_id IN (SELECT id FROM sequence_instance WHERE retain_until < ?1)")
-    void deleteExpired(ZonedDateTime cutoffTime);
 
     @Modifying
     @Query(nativeQuery = true, value = "DELETE FROM buffered_message WHERE sequence_instance_id IN (SELECT id FROM sequence_instance WHERE state = 'CLOSED')")
@@ -21,4 +15,18 @@ interface SpringDataJpaBufferedMessageRepository extends JpaRepository<BufferedM
 
     @Query("from BufferedMessage bm left join fetch bm.headers where bm.sequencedMessageId = :sequencedMessageId")
     BufferedMessage getBySequencedMessageId(long sequencedMessageId);
+
+    @Modifying
+    @Query(
+            nativeQuery = true,
+            value = """
+                DELETE FROM buffered_message
+                WHERE sequence_instance_id = ?1
+                  AND EXISTS (
+                      SELECT 1 FROM sequence_instance WHERE id = ?1 AND state <> 'CLOSED'
+                )
+                """
+    )
+    int deleteForNotClosedSequencedId(long sequenceInstanceId);
+
 }